harden: comprehensive security pattern enforcement across codebase
Apply hardening patterns from MEMORY.md across 100+ files:
- Fix fail-closed deadline semantics in obligation_channel.rs
- Replace unsafe casts with try_from() in cryptographic hash functions
- Implement saturating arithmetic for all counter operations
- Standardize push_bounded pattern for Vec capacity management
- Add length validation in canonical serialization functions

Security impact: prevents overflow attacks, timing side-channels,
hash collisions, and ensures fail-closed boundary semantics.

+++ crates/franken-node/src/remote/mod.rs
@@ -703,7 +703,7 @@ mod remote_module_negative_tests {
 
             // Test corruption bit count boundaries
             FaultConfig { corrupt_bit_count: 0, ..valid_config() },
-            FaultConfig { corrupt_bit_count: u32::MAX, ..valid_config() },
+            FaultConfig { corrupt_bit_count: usize::MAX, ..valid_config() },
         ];
 
         for (i, config) in boundary_configs.iter().enumerate() {
